requirements.txt
plotly>=2.4.1
PyYAML>=3.12
